requirements.txt
flask
python-dotenv
alpaca-trade-api
langchain
langchain-google-genai
google-generativeai
